backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
PyMuPDF>=1.23.4
Pillow>=10.0.1
aiofiles>=23.2.1
pytesseract>=0.3.10
opencv-python-headless>=4.8.1.78
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.26.0
